DATA_CHUNK hook: reset timeout on every read
I have encountered situations in the wild where someone was trying to
send a large message over a slow connection that would require a very
large static timeout (>hour) to accomodate. This way you can keep the
timeout reasonably short (minutes) and it can take as long as it takes
as long as the client keeps making forward progress.

We could make this available to the current api under control of a
flag but I don't want to unexpectedly change the behavior for
exiseting users.